+++ packages/stream_chat_flutter/lib/src/components/message_composer/message_composer_recording_ongoing.dart
@@ -50,7 +50,7 @@ class StreamMessageComposerRecordingOngoing extends StatelessWidget {
           ),
           const SizedBox(width: 23),
           _GradientText(
-            'Slide to cancel',
+            context.translations.slideToCancelLabel,
             style: context.streamTextTheme.bodyDefault,
             gradient: LinearGradient(
               colors: [colorScheme.textPrimary, colorScheme.textTertiary],